+++ src/main/java/r2graph/io/MappingDocument.java
@@ -1,4 +1,4 @@
-package r2graph.util;
+package r2graph.io;
 
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.ModelFactory;
