+++ src/test/java/r2graph/r2rml/R2RMLParserTest.java
@@ -11,7 +11,7 @@ import r2graph.configmap.ConfigMap;
 import r2graph.configmap.EntityMap;
 import r2graph.configmap.PredicateObjectMap;
 import r2graph.configmap.impl.PredicateObjectMapCom;
-import r2graph.util.MappingDocument;
+import r2graph.io.MappingDocument;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
